requirements.txt
# Core EEG processing
mne>=1.6.0
pyedflib>=0.1.30
wfdb>=4.1.0

# Numerical / data
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Deep learning
torch>=2.1.0
torchvision>=0.16.0

# ML utilities
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# LLM (Baseline 2)
openai>=1.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
jupyterlab>=4.0.0
